network: Consolidate utility functions
