+++ lxd/networks.go
@@ -8,7 +8,6 @@ import (
 	"net/http"
 	"os"
 	"os/exec"
-	"regexp"
 	"strconv"
 
 	"github.com/gorilla/mux"
@@ -59,25 +58,6 @@ func networkIsInUse(c container, name string) bool {
 	return false
 }
 
-func networkValidateName(name string) error {
-	// Validate the length
-	if len(name) < 2 {
-		return fmt.Errorf("Interface name is too short (minimum 2 characters)")
-	}
-
-	if len(name) > 15 {
-		return fmt.Errorf("Interface name is too long (maximum 15 characters)")
-	}
-
-	// Validate the character set
-	match, _ := regexp.MatchString("^[-a-zA-Z0-9]*$", name)
-	if !match {
-		return fmt.Errorf("Interface name contains invalid characters")
-	}
-
-	return nil
-}
-
 // API endpoints
 func networksGet(d *Daemon, r *http.Request) Response {
 	recursionStr := r.FormValue("recursion")
@@ -126,7 +106,7 @@ func networksPost(d *Daemon, r *http.Request) Response {
 		return BadRequest(fmt.Errorf("No name provided"))
 	}
 
-	err = networkValidateName(req.Name)
+	err = networkValidName(req.Name)
 	if err != nil {
 		return BadRequest(err)
 	}
@@ -305,7 +285,7 @@ func networkPost(d *Daemon, r *http.Request) Response {
 		return BadRequest(fmt.Errorf("No name provided"))
 	}
 
-	err = networkValidateName(req.Name)
+	err = networkValidName(req.Name)
 	if err != nil {
 		return BadRequest(err)
 	}
